# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Helpdesk API Documentation",
        "description": "Dokumentasi API untuk aplikasi E-Ticketing Helpdesk berbasis Laravel.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/tickets/{ticketId}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Menampilkan komentar tiket",
                "description": "Mengambil daftar komentar berdasarkan ID tiket.",
                "operationId": "967dbbe59bd1836606fc9b462f873648",
                "parameters": [
                    {
                        "name": "ticketId",
                        "in": "path",
                        "description": "ID tiket",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Komentar berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Tiket tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Menambahkan komentar tiket",
                "description": "Menambahkan komentar baru pada tiket tertentu.",
                "operationId": "55b06c1a998d64bf36d42c1ef5ea923c",
                "parameters": [
                    {
                        "name": "ticketId",
                        "in": "path",
                        "description": "ID tiket",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Saya sudah mencoba restart laptop, tapi masih tidak menyala."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Komentar berhasil ditambahkan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "404": {
                        "description": "Tiket tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comments/{id}": {
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Menghapus komentar",
                "description": "Menghapus komentar berdasarkan ID. Biasanya hanya admin yang diperbolehkan.",
                "operationId": "b683c30a239aa780ffc7e9b219e90811",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID komentar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Komentar berhasil dihapus"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Tidak memiliki akses"
                    },
                    "404": {
                        "description": "Komentar tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/stats": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Menampilkan statistik dashboard",
                "description": "Mengambil ringkasan statistik tiket seperti total tiket, tiket open, in progress, resolved, dan closed.",
                "operationId": "654e1617f74d2b9e0654ec77620d960b",
                "responses": {
                    "200": {
                        "description": "Statistik dashboard berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/recent-tickets": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Menampilkan tiket terbaru",
                "description": "Mengambil daftar tiket terbaru untuk ditampilkan pada dashboard.",
                "operationId": "5078f06fce8d9b2a0fddc4012f820b4f",
                "responses": {
                    "200": {
                        "description": "Tiket terbaru berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}